+++ src/Vector/BLF/CanFdErrorFrame64.cpp
@@ -52,13 +52,14 @@ void CanFdErrorFrame64::read(AbstractFile & is) {
     is.read(reinterpret_cast<char *>(data.data()), static_cast<std::streamsize>(data.size()));
     if (hasExtData())
         CanFdExtFrameData::read(is);
-    reservedCanFdErrorFrame64.resize(objectSize - calculateObjectSize());
-    is.read(reinterpret_cast<char *>(reservedCanFdErrorFrame64.data()), static_cast<std::streamsize>(reservedCanFdErrorFrame64.size()));
+    // @note reservedCanFdExtFrameData is read here as CanFdExtFrameData doesn't know the objectSize
+    reservedCanFdExtFrameData.resize(objectSize - calculateObjectSize());
+    is.read(reinterpret_cast<char *>(reservedCanFdExtFrameData.data()), static_cast<std::streamsize>(reservedCanFdExtFrameData.size()));
 }
 
 void CanFdErrorFrame64::write(AbstractFile & os) {
     /* pre processing */
-    validDataBytes = static_cast<uint16_t>(data.size());
+    validDataBytes = static_cast<uint8_t>(data.size());
 
     ObjectHeader::write(os);
     os.write(reinterpret_cast<char *>(&channel), sizeof(channel));
@@ -82,7 +83,6 @@ void CanFdErrorFrame64::write(AbstractFile & os) {
     os.write(reinterpret_cast<char *>(data.data()), static_cast<std::streamsize>(data.size()));
     if (hasExtData())
         CanFdExtFrameData::write(os);
-    os.write(reinterpret_cast<char *>(reservedCanFdErrorFrame64.data()), static_cast<std::streamsize>(reservedCanFdErrorFrame64.size()));
 }
 
 bool CanFdErrorFrame64::hasExtData() const {
@@ -115,7 +115,6 @@ uint32_t CanFdErrorFrame64::calculateObjectSize() const {
         static_cast<uint32_t>(data.size());
     if (hasExtData())
         size += CanFdExtFrameData::calculateObjectSize();
-    size += static_cast<uint32_t>(reservedCanFdErrorFrame64.size());
     return size;
 }