+++ src/Vector/BLF/FileStatistics.cpp
@@ -45,7 +45,7 @@ void FileStatistics::read(AbstractFile & is) {
     is.read(reinterpret_cast<char *>(&applicationBuild), sizeof(applicationBuild));
     is.read(reinterpret_cast<char *>(&measurementStartTime), sizeof(measurementStartTime));
     is.read(reinterpret_cast<char *>(&lastObjectTime), sizeof(lastObjectTime));
-    is.read(reinterpret_cast<char *>(&fileSizeWithoutUnknown115), sizeof(fileSizeWithoutUnknown115));
+    is.read(reinterpret_cast<char *>(&restorePointsOffset), sizeof(restorePointsOffset));
     is.read(reinterpret_cast<char *>(reservedFileStatistics.data()), static_cast<std::streamsize>(reservedFileStatistics.size() * sizeof(uint32_t)));
 }
 
@@ -63,7 +63,7 @@ void FileStatistics::write(AbstractFile & os) {
     os.write(reinterpret_cast<char *>(&applicationBuild), sizeof(applicationBuild));
     os.write(reinterpret_cast<char *>(&measurementStartTime), sizeof(measurementStartTime));
     os.write(reinterpret_cast<char *>(&lastObjectTime), sizeof(lastObjectTime));
-    os.write(reinterpret_cast<char *>(&fileSizeWithoutUnknown115), sizeof(fileSizeWithoutUnknown115));
+    os.write(reinterpret_cast<char *>(&restorePointsOffset), sizeof(restorePointsOffset));
     os.write(reinterpret_cast<char *>(reservedFileStatistics.data()), static_cast<std::streamsize>(reservedFileStatistics.size() * sizeof(uint32_t)));
 }
 
@@ -82,7 +82,7 @@ uint32_t FileStatistics::calculateStatisticsSize() const {
         sizeof(applicationBuild) +
         sizeof(measurementStartTime) +
         sizeof(lastObjectTime) +
-        sizeof(fileSizeWithoutUnknown115) +
+        sizeof(restorePointsOffset) +
         static_cast<uint32_t>(reservedFileStatistics.size() * sizeof(uint32_t));
 }