+++ CHANGELOG.md
@@ -3,14 +3,22 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Drop CMAKE_BUILD_TYPE from CMakeLists to allow passing it to cmake.
+- CanFdMessage handling. ValidDataBytes set automatically.
+### Fixed
+- Fix ObjectQueue virtuals. Clean queue in destructor.
+
+## [2.4.0] - 2021-01-29
 ### Added
 - New Types: CanSettingChanged, DistributedObjectMember, AttributeEvent
+- Add support for Restore Points (which are stored in Unknown115 objects)
 ### Changed
 - Drop VectorTypes.h and use std types instead.
 ### Fixed
 - Fixes in CanErrorFrameExt, CanFdMessage, CanFdMessage64, CanMessage2, EthernetFrame, LinMessage2
 - Fix potential deadlock in File::close
-- Fix FileStatistics: application, apiNumber, compressionLevel, objectsRead.
+- Fix FileStatistics: application, apiNumber, compressionLevel, objectsRead
 
 ## [2.3.1] - 2020-07-16
 ### Fixed